pyproject.toml
[project]
name = "rkexamspiders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "curl-cffi>=0.14.0",
    "dbutils>=3.1.2",
    "ddddocr>=1.6.1",
    "pycryptodome>=3.23.0",
    "pymysql>=1.1.2",
    "pyyaml>=6.0.3",
    "scrapy>=2.14.2",
]
